fix(#321, #317): /audit 500 — Option<u64> vs CAST AS SIGNED type mismatch (#322)
`src/routes/audit.rs:129` projects `CAST(v.location_id AS SIGNED)
AS location_id` (SQL type `BIGINT`) but the tuple decoder typed
it as `Option<u64>` (SQL type `BIGINT UNSIGNED`). sqlx 0.8 refuses
the mismatch — textbook violation of `CLAUDE.md` MariaDB gotcha
#2 ("use `Option<i64>`, then convert").

The bug was latent since v1.6.0 (#237 shelf-audit) because the
empty-result branch never decoded a row and the CI E2E coverage
seeded zero flagged volumes. Surfaced in prod 2026-05-22 the
moment the user actually flagged a volume "À contrôler" on the
v1.7.2 NAS — the /audit page returned 500 with the browser's
default error icon, which is also what produced the mystery
"large black circle with cross" reported earlier as #317.

Fix: decode as `Option<i64>`, convert `as u64` when feeding it
into `LocationModel::get_path(pool, u64)`. Mirrors the existing
idiom in `src/models/location.rs::get_path` line 133.

Regression test `tests/audit_list_non_empty.rs` seeds a flagged
volume + storage_location + title and asserts `GET /audit`
returns 200 with the rendered HTML containing both the volume
label and the location-path segment — exactly the path that
crashed in prod. Also keeps a companion empty-state assertion to
lock the happy-path invariant.

Bumps Cargo.toml + Cargo.lock + manual cover (EN + FR) → 1.7.3.
New "What's new in 1.7.3" section in chapter 8 of the manual,
EN + FR; PDFs rebuilt.

Closes #321, closes #317.

+++ src/routes/audit.rs
@@ -126,7 +126,16 @@ pub async fn audit_list_page(
     // sorted location → V-code so the user walks the shelf in order.
     // The location join is LEFT — an unshelved flagged volume still
     // surfaces, just with an empty path.
-    let rows = sqlx::query_as::<_, (u64, String, u64, String, Option<u64>)>(
+    //
+    // Fix #321 — location_id is `BIGINT UNSIGNED NULL` but we project
+    // it as `CAST(... AS SIGNED)`. Per CLAUDE.md MariaDB type gotcha
+    // #2 the decoder MUST be `Option<i64>`, then convert to `u64` for
+    // the downstream `LocationModel::get_path(pool, u64)` call. The
+    // original `Option<u64>` triggered a sqlx 0.8 type-mismatch and a
+    // 500 the moment a flagged-volume row actually existed in prod.
+    // Latent since v1.6.0 (#237) — covered by the empty-state branch
+    // in CI but never exercised on a non-empty result set.
+    let rows = sqlx::query_as::<_, (u64, String, u64, String, Option<i64>)>(
         "SELECT v.id, v.label, t.id, t.title, CAST(v.location_id AS SIGNED) AS location_id \
          FROM volumes v \
          JOIN titles t ON t.id = v.title_id AND t.deleted_at IS NULL \
@@ -139,7 +148,7 @@ pub async fn audit_list_page(
     let mut items: Vec<AuditRowDisplay> = Vec::with_capacity(rows.len());
     for (vol_id, vol_label, title_id, title_name, location_id) in rows {
         let location_path = match location_id {
-            Some(lid) => crate::models::location::LocationModel::get_path(pool, lid)
+            Some(lid) => crate::models::location::LocationModel::get_path(pool, lid as u64)
                 .await
                 .unwrap_or_default(),
             None => String::new(),
